chore(locale): update no.json
The `settings` locale is for some reason not included in most (if not all) non-english locale files. I'm not sure if that is intentional or not, so please let me know if you want that removed again.

The reason for changing `cancel_progress` is that "progresjonsbar" seems to be the industry standard translation for "progress bar", it also sounds a lot less clunky than "fremdriftsindikatoren".