Allow explicit rotation order for Progress props
As the rotation order varies wildly across community scripts, and Rockstar usually uses order 2, allowing overrides here is important for compatibility when updating legacy resources.

From a personal perspective, order 0 is kinda just a pain in the ass to use. The X and Z axes have to be manipulated in inverse to each other to produce meaningful rotation. There is also an inversion zone, which causes unexpected shifting at some points of X/Z crossover. Order 1, as well as some others, are much more friendly to work with.